UISS-189: Switch from Djatoka to IIIF-compatible image server
In the process of switching to IIIF compatible server, we've also
removed support for multiple resource URIs / Tile Sources due to the
complexity around handling multiple asynchronous json requests to obtain
the width/height definition of each image.

We suspect that this feature was not being used at all in the first
place.